+++ src/main/kotlin/com/autonomousapps/advice/PluginAdvice.kt
@@ -1,15 +1,22 @@
 package com.autonomousapps.advice
 
+import org.gradle.api.Incubating
+
+@Incubating
 data class PluginAdvice(
   val redundantPlugin: String,
   val reason: String
 ) : Comparable<PluginAdvice> {
 
   companion object {
-    // TODO make more flexible
-    fun redundantPlugin() = PluginAdvice(
+    fun redundantJavaLibrary() = PluginAdvice(
       redundantPlugin = "java-library",
-      reason = "This project has both java-library and org.jetbrains.kotlin.jvm applied, which is redundant. You can remove java-library"
+      reason = "this project has both java-library and org.jetbrains.kotlin.jvm applied, which is redundant. You can remove java-library"
+    )
+
+    fun redundantKotlinJvm() = PluginAdvice(
+      redundantPlugin = "org.jetbrains.kotlin.jvm",
+      reason = "this project has both java-library and org.jetbrains.kotlin.jvm applied, which is redundant. You can remove org.jetbrains.kotlin.jvm"
     )
   }